Refactor to use the new Venue page from Untappd
This removes the search query that was going haywire.

Unfortunately we still have to page-scrape to get the listing
for beers on draft, since there's no public API for that which
is pretty lame.

At least there is always an associated Untappd beer entry now,
and there will be no more false-positives due to the search
query being incorrect.

Also now uses `async`/`await` instead of `yield` and "suspend" module